ai-service/requirements.txt
flask==2.0.1
numpy
pillow
face-recognition 